Add custom style
