+++ integracao-rd-station.php
@@ -77,14 +77,5 @@ function rdsm_tracking_code_hooks() {
 $rdsm_tracking_code_hook = new RDSMTrackingCodeHooks;
 $rdsm_tracking_code_hook->handle();
 
-add_action( 'admin_enqueue_scripts', 'enqueue_rd_admin_style' );
-function enqueue_rd_admin_style($hook) {
-  $screen = get_current_screen();
 
-  if ($screen->base === 'settings_page_rdstation-settings-page') {
-    wp_enqueue_script( 'rd_admin_script', plugin_dir_url( __FILE__ ) . 'assets/js/admin.js' );
-  }
-
-  if ( 'post.php' != $hook ) return;
-  wp_enqueue_style( 'rd_admin_style', plugin_dir_url( __FILE__ ) . 'assets/styles/admin.css' );
-}
+require_once("load_scripts.php");
