+++ settings/settings_menu.php
@@ -2,7 +2,9 @@
 
 add_action( 'admin_menu', 'rdstation_menu' );
 function rdstation_menu() {
-  add_options_page(
+  global $rdsm_settings_page;
+
+  $rdsm_settings_page = add_options_page(
     __('RD Station Settings', 'integracao-rd-station'),
     __('RD Station Settings', 'integracao-rd-station'),
     'manage_options',
